pyproject.toml
[project]
name = "ml-project-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3",
    "fastapi",
    "ipykernel>=7.2.0",
    "joblib",
    "lightning>=2.6.1",
    "mlflow>=3.9.0",
    "numpy<2; sys_platform == 'darwin'",
    "numpy; sys_platform == 'linux'",
    "pandas",
    "python-dotenv",
    "pip>=26.0.1",
    "s3fs>=2026",
    "scikit-learn",
    "torch>=2.0,<2.3; sys_platform == 'darwin'",
    "torch>=2.0; sys_platform == 'linux'",
    "uvicorn[standard]",
]

[project.scripts]
ml-project-template = "ml_project_template:main"

[dependency-groups]
dev = ["httpx>=0.27", "pytest>=8.0"]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
